refactor: extract inline test modules to sibling files (under 1000-line cap)
Five production files sat in the 930-984 line danger zone of the 1000-line
file-size CI cap, each ending in a large inline #[cfg(test)] mod tests block.
Move each block verbatim into a sibling test file, leaving a module declaration:

- crates/astrid-core/src/dirs.rs -> dirs_tests.rs
- crates/astrid-storage/src/secret.rs -> secret_tests.rs
- crates/astrid-capsule/src/security/manifest_gate.rs -> manifest_gate_tests.rs
- crates/astrid-approval/src/policy.rs -> policy_tests.rs
- crates/astrid-kernel/src/kernel_router/mod.rs -> kernel_router/tests.rs

Pure mechanical relocation: no behaviour, signature, or visibility change.

Closes #956.
