+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 
 ### Changed
 
+- Extracted the inline `#[cfg(test)] mod tests` blocks from `dirs.rs`, `secret.rs`, `kernel_router/mod.rs`, `manifest_gate.rs`, and `policy.rs` into sibling test files to keep each production file under the 1000-line file-size cap; pure mechanical move, no behaviour change. Closes #956.
+
 - **CI `Test` job strips debug info so linking the `astrid-capsule` test binary no longer exhausts the runner.** The `astrid-capsule` test binary statically links the entire dependency tree (wasmtime, aws-lc, ring, surrealkv, …); after a GitHub `ubuntu-latest` runner-image rotation the final link began dying with `collect2: fatal error: ld terminated with signal 7 [Bus error]` (an OOM/disk-pressure failure), turning every `Test (ubuntu-latest)` run red and cancelling `macos-latest` via matrix fail-fast — independent of the change under test (it reproduced on a main commit that touched neither capsule code). The Test job now sets `CARGO_PROFILE_DEV_DEBUG=0` / `CARGO_PROFILE_TEST_DEBUG=0`, dropping debug info (the dominant contributor to both linker peak memory and `target/` size) for the test build only. CI-scoped via Cargo profile env overrides — local and dev builds keep full debug info. Closes #954.
 
 - **Manual `astrid capsule install @org/repo` against a multi-capsule release now installs every capsule the release ships by default, with `--capsule <name>` to pick one.** #921/#922 threaded the distro capsule `name` into `pick_capsule` so a monorepo release could select the right archive, but left the *manual* install path with no selector: a release shipping several `.capsule` assets passed no hint and hit `pick_capsule`'s "no capsule name to pick one" hard error, so there was no way to install such a release by hand at all. The manual path now defaults to installing **all** capsules in the release — `install_from_github` collects every `.capsule` asset (via a new pure `capsule_assets` helper over the release JSON) and, when no selector is given, downloads and unpacks each one. Installation is **best-effort**: it prints `Installing <name>...` per archive, reports any that fail and keeps going, then returns an error naming all failures if any did — never silently swallowing a failed install. A new `--capsule <NAME>` flag opts back into single-capsule selection (`<name>.capsule` via the unchanged `pick_capsule`), and a single-asset release installs that one capsule exactly as before. The distro path (`install_capsule_batch`) is **unchanged** — it keeps selecting by the distro capsule `name`. The `update` path now passes the capsule being updated as the selector so a monorepo source refreshes only that one capsule, not the whole release. The clone+build fallback (no release assets) stays single-capsule; making `astrid-build` handle a virtual-workspace root is a separate gap. Closes #925.
