+++ crates/astrid-kernel/src/kernel_router/tests.rs
@@ -1,0 +1,248 @@
+//! Tests for `kernel_router/mod.rs`. Split out to keep `mod.rs` under the
+//! 1000-line CI threshold. Included as a `tests` submodule of `kernel_router`.
+
+use super::*;
+
+#[test]
+fn rate_limiter_allows_within_limit() {
+    let mut limiter = ManagementRateLimiter::new();
+    for _ in 0..5 {
+        assert!(limiter.check("ReloadCapsules", 5));
+    }
+    // 6th should be rejected
+    assert!(!limiter.check("ReloadCapsules", 5));
+}
+
+#[test]
+fn rate_limiter_independent_buckets() {
+    let mut limiter = ManagementRateLimiter::new();
+    // Fill ReloadCapsules
+    for _ in 0..5 {
+        assert!(limiter.check("ReloadCapsules", 5));
+    }
+    assert!(!limiter.check("ReloadCapsules", 5));
+
+    // InstallCapsule should still be allowed
+    assert!(limiter.check("InstallCapsule", 10));
+}
+
+#[test]
+fn rate_limiter_sliding_window_eviction() {
+    let mut limiter = ManagementRateLimiter::new();
+    // Fill the bucket
+    for _ in 0..5 {
+        assert!(limiter.check("ReloadCapsules", 5));
+    }
+    assert!(!limiter.check("ReloadCapsules", 5));
+
+    // Manually set all timestamps to 61 seconds ago to simulate expiry.
+    if let Some(timestamps) = limiter.buckets.get_mut("ReloadCapsules") {
+        let past = Instant::now() - std::time::Duration::from_secs(61);
+        for ts in timestamps.iter_mut() {
+            *ts = past;
+        }
+    }
+
+    // Should be allowed again after old entries are evicted
+    assert!(limiter.check("ReloadCapsules", 5));
+}
+
+#[test]
+fn rate_limiter_sliding_window_prevents_boundary_burst() {
+    let mut limiter = ManagementRateLimiter::new();
+    // Fill 5 requests
+    for _ in 0..5 {
+        assert!(limiter.check("ReloadCapsules", 5));
+    }
+
+    // Move only 3 of the 5 timestamps to the past (beyond 60s window).
+    // This simulates partial window expiry - only 3 slots should free up.
+    if let Some(timestamps) = limiter.buckets.get_mut("ReloadCapsules") {
+        let past = Instant::now() - std::time::Duration::from_secs(61);
+        for ts in timestamps.iter_mut().take(3) {
+            *ts = past;
+        }
+    }
+
+    // Should allow exactly 3 more (the evicted slots), not 5
+    for _ in 0..3 {
+        assert!(limiter.check("ReloadCapsules", 5));
+    }
+    assert!(!limiter.check("ReloadCapsules", 5));
+}
+
+#[test]
+fn rate_limit_for_request_returns_correct_limits() {
+    let (name, limit) = rate_limit_for_request(&KernelRequest::ReloadCapsules);
+    assert_eq!(name, "ReloadCapsules");
+    assert_eq!(limit, Some(5));
+
+    let (name, limit) = rate_limit_for_request(&KernelRequest::ListCapsules);
+    assert_eq!(name, "ListCapsules");
+    assert_eq!(limit, None);
+}
+
+// ── Capability mapping (issue #670) ──────────────────────────────
+
+fn all_request_variants() -> Vec<KernelRequest> {
+    vec![
+        KernelRequest::Shutdown { reason: None },
+        KernelRequest::GetStatus,
+        KernelRequest::ReloadCapsules,
+        KernelRequest::InstallCapsule {
+            source: "x".to_string(),
+            workspace: false,
+        },
+        KernelRequest::ListCapsules,
+        KernelRequest::GetCommands,
+        KernelRequest::GetCapsuleMetadata,
+        KernelRequest::ApproveCapability {
+            request_id: "r".to_string(),
+            signature: "s".to_string(),
+        },
+    ]
+}
+
+#[test]
+fn required_capability_every_variant_has_non_empty_mapping() {
+    for req in all_request_variants() {
+        let cap = required_capability(&req, AuthorityScope::Self_);
+        assert!(
+            !cap.is_empty(),
+            "required_capability returned empty for {req:?}"
+        );
+    }
+}
+
+#[test]
+fn required_capability_mapping_per_variant_self_scope() {
+    assert_eq!(
+        required_capability(
+            &KernelRequest::Shutdown { reason: None },
+            AuthorityScope::Self_
+        ),
+        "system:shutdown"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::GetStatus, AuthorityScope::Self_),
+        "system:status"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::ReloadCapsules, AuthorityScope::Self_),
+        "self:capsule:reload"
+    );
+    assert_eq!(
+        required_capability(
+            &KernelRequest::InstallCapsule {
+                source: String::new(),
+                workspace: false
+            },
+            AuthorityScope::Self_
+        ),
+        "self:capsule:install"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::ListCapsules, AuthorityScope::Self_),
+        "self:capsule:list"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::GetCommands, AuthorityScope::Self_),
+        "self:capsule:list"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::GetCapsuleMetadata, AuthorityScope::Self_),
+        "self:capsule:list"
+    );
+    assert_eq!(
+        required_capability(
+            &KernelRequest::ApproveCapability {
+                request_id: String::new(),
+                signature: String::new(),
+            },
+            AuthorityScope::Self_
+        ),
+        "self:approval:respond"
+    );
+}
+
+#[test]
+fn required_capability_mapping_global_scope() {
+    // Global scope strips the `self:` prefix from capsule operations
+    // (Layer 6 will start using this when cross-agent variants land).
+    assert_eq!(
+        required_capability(&KernelRequest::ReloadCapsules, AuthorityScope::Global),
+        "capsule:reload"
+    );
+    assert_eq!(
+        required_capability(
+            &KernelRequest::InstallCapsule {
+                source: String::new(),
+                workspace: false
+            },
+            AuthorityScope::Global
+        ),
+        "capsule:install"
+    );
+    assert_eq!(
+        required_capability(&KernelRequest::ListCapsules, AuthorityScope::Global),
+        "capsule:list"
+    );
+    // system:* variants are scope-invariant.
+    assert_eq!(
+        required_capability(
+            &KernelRequest::Shutdown { reason: None },
+            AuthorityScope::Global
+        ),
+        "system:shutdown"
+    );
+}
+
+#[test]
+fn resolve_scope_defaults_to_self() {
+    let caller = PrincipalId::new("alice").unwrap();
+    for req in all_request_variants() {
+        assert_eq!(
+            resolve_scope(&req, &caller),
+            AuthorityScope::Self_,
+            "scope should default to Self_ for today's variants"
+        );
+    }
+}
+
+// ── Caller resolution ────────────────────────────────────────────
+
+#[test]
+fn resolve_caller_uses_ipc_principal_when_present() {
+    let mut msg = IpcMessage::new(
+        "astrid.v1.request.system",
+        IpcPayload::RawJson(serde_json::json!({})),
+        uuid::Uuid::nil(),
+    );
+    msg.principal = Some("alice".to_string());
+    let caller = resolve_caller(&msg);
+    assert_eq!(caller.as_str(), "alice");
+}
+
+#[test]
+fn resolve_caller_falls_back_to_default_when_missing() {
+    let msg = IpcMessage::new(
+        "astrid.v1.request.system",
+        IpcPayload::RawJson(serde_json::json!({})),
+        uuid::Uuid::nil(),
+    );
+    let caller = resolve_caller(&msg);
+    assert_eq!(caller, PrincipalId::default());
+}
+
+#[test]
+fn resolve_caller_falls_back_to_default_on_invalid_principal() {
+    let mut msg = IpcMessage::new(
+        "astrid.v1.request.system",
+        IpcPayload::RawJson(serde_json::json!({})),
+        uuid::Uuid::nil(),
+    );
+    // Invalid principal chars → PrincipalId::new fails → fall back.
+    msg.principal = Some("alice@evil.example".to_string());
+    let caller = resolve_caller(&msg);
+    assert_eq!(caller, PrincipalId::default());
+}
